feat(#431): apply implement pass changes
